requirements.txt
html2image==2.0.4
fastapi==0.115.0
uvicorn[standard]==0.30.0
Pillow==10.4.0
playwright==1.40.0
